requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
attrs=23.2.0=pypi_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
dataclasses-json=0.6.5=pypi_0
distro=1.9.0=pypi_0
frozenlist=1.4.1=pypi_0
groq=0.5.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
langchain=0.1.16=pypi_0
langchain-community=0.0.34=pypi_0
langchain-core=0.1.46=pypi_0
langchain-groq=0.1.3=pypi_0
langchain-openai=0.1.4=pypi_0
langchain-text-splitters=0.0.1=pypi_0
langgraph=0.0.39=pypi_0
langsmith=0.1.51=pypi_0
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.45.3=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
loguru=0.7.2=pypi_0
marshmallow=3.21.1=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4.20240210=h078ce10_0
numpy=1.26.4=pypi_0
openai=1.23.6=pypi_0
openssl=3.2.1=h0d3ecfb_1
orjson=3.10.1=pypi_0
outcome=1.3.0.post0=pypi_0
packaging=23.2=pypi_0
pip=24.0=pyhd8ed1ab_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pyotp=2.9.0=pypi_0
pysocks=1.7.1=pypi_0
python=3.12.3=h4a7b5fc_0_cpython
python-dotenv=1.0.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h92ec313_1
regex=2024.4.16=pypi_0
requests=2.31.0=pypi_0
selenium=4.20.0=pypi_0
setuptools=69.5.1=pyhd8ed1ab_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlalchemy=2.0.29=pypi_0
tenacity=8.2.3=pypi_0
tiktoken=0.6.0=pypi_0
tk=8.6.13=h5083fa2_1
todoist-api-python=2.1.3=pypi_0
tqdm=4.66.2=pypi_0
trio=0.25.0=pypi_0
trio-websocket=0.11.1=pypi_0
typing-extensions=4.11.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024a=h0c530f3_0
urllib3=2.2.1=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
wsproto=1.2.0=pypi_0
xz=5.2.6=h57fd34a_0
yarl=1.9.4=pypi_0
